Fix UI5 MCP server logo
Somehow the clip path inside the SVG logo prevented it from being
displayed. This change removes the (anyway invisible) inner structure
and only maintains the outline, fixing the display in the web page. It
still uses "currentColor" as color, so the icon uses the locally valid
text color from CSS.